// src/open-api/model/bulkStatusUpdateCommand.ts
/**
 * Receipt Wrangler API.
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 5.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BulkStatusUpdateCommand { 
    /**
     * Optional comment to leave on each receipt
     */
    comment?: string;
    /**
     * Status to update to
     */
    status: string;
    /**
     * Receipt ids to update
     */
    receiptIds: Array<number>;
}