// src/open-api/model/category.ts
/**
 * Receipt Wrangler API.
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Category to relate receipts to
 */
export interface Category { 
    createdAt?: string;
    createdBy?: number;
    id?: number;
    /**
     * Name of the category
     */
    name?: string;
    /**
     * Description of the category
     */
    description?: string;
    updatedAt?: string;
}